renames detector module
