+++ src/lib/designAntipatternDetectors.ts
